+++ src/Main/ProjectExplorerPage.py
@@ -87,7 +87,7 @@ class ProjectExplorerPage(QtWidgets.QWidget, NavigationPage):
         self._toolbar.setContextMenuPolicy(QtCore.Qt.PreventContextMenu)  # Disable right click menu (wtf Qt?)
         self._toolbar.setIconSize(QtCore.QSize(32, 32))
         self._toolbar.setToolButtonStyle(QtCore.Qt.ToolButtonStyle.ToolButtonTextBesideIcon)
-        self._toolbar.setStyleSheet("""QToolBar QToolButton { width: 180px; }""")
+        self._toolbar.setStyleSheet("""QToolBar QToolButton { width: 200px; }""")
         self._toolbar.setOrientation(QtCore.Qt.Orientation.Vertical)
 
         # self.addToolBar(QtCore.Qt.ToolBarArea.LeftToolBarArea, self._toolbar)
@@ -207,7 +207,11 @@ class ProjectExplorerPage(QtWidgets.QWidget, NavigationPage):
                 __("No faces found in the project. Please add images with faces to the project."))
             return
 
-        self._shell.openPage(GroupFacesPage())
+        pageIndex = self._shell.pageIndex(GroupFacesPage)
+        if pageIndex >= 0:
+            self._shell.displayPageAtIndex(pageIndex)
+        else:
+            self._shell.openPage(GroupFacesPage())
 
     @QtCore.Slot(list)
     def _onImagesAdded(self, images: list[Image]) -> None: